Extracted common active effect summarization code to share.
