+++ src/components/item-list/EffectSummary.svelte
@@ -3,6 +3,7 @@
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
   import type { ActiveEffect5e } from 'src/types/types';
   import HorizontalLineSeparator from '../layout/HorizontalLineSeparator.svelte';
+  import { ActiveEffectsHelper } from 'src/utils/active-effect';
 
   interface Props {
     activeEffect: ActiveEffect5e;
@@ -14,43 +15,9 @@
     FoundryAdapter.enrichHtml(activeEffect.description ?? ''),
   );
 
-  let pills = $derived.by(() => {
-    let result = [];
-
-    if (activeEffect.disabled) {
-      result.push('EFFECT.Disabled');
-    }
-
-    if (activeEffect.transfer) {
-      result.push('EFFECT.Transfer');
-    }
-
-    if (activeEffect.isSuppressed) {
-      result.push('DND5E.Suppressed');
-    }
-
-    Array.from<string>(activeEffect.statuses)
-      .map(
-        (x: string) => CONFIG.statusEffects.find((y) => y.id === x)?.name ?? x,
-      )
-      .forEach((e) => {
-        result.push(e);
-      });
-
-    return result;
-  });
-
-  function findMode(mode: number) {
-    const entry = Object.entries(CONST.ACTIVE_EFFECT_MODES).find(
-      ([_, value]) => value === mode,
-    );
-
-    if (!entry) {
-      return '—';
-    }
-
-    return localize(`EFFECT.MODE_${entry[0]}`);
-  }
+  let pills = $derived.by(() =>
+    ActiveEffectsHelper.getActiveEffectPills(activeEffect),
+  );
 
   const localize = FoundryAdapter.localize;
 </script>
@@ -84,7 +51,7 @@
         </thead>
         <tbody>
           {#each activeEffect.changes as change}
-            {@const modeLabel = findMode(change.mode)}
+            {@const modeLabel = ActiveEffectsHelper.findMode(change.mode)}
 
             <tr>
               <td
